+++ src/database/datatype.rs
@@ -8,7 +8,7 @@ pub enum DataType {
     TinyInt { x: i8 },
     Decimal { x: f32 },
     Float { x: f64 },
-    VarChar { x: String, y: usize },
+    VarChar { x: String, y: u16 },
     Bool { x: bool },
     Date { x: i64 },
     Time { x: i64 },
